* ext/oci8/oci8.c: Add a workaround for Oracle 9.2.0.1 when clearing     a client identifier. * ext/oci8/oci8lib.c: Fix a segv when canceling a non-blocking     execution. * test_appinfo.rb: Fix temporarily to prevent a segmentation fault     under Oracle 9.2.0.1. * test/test_oci8.rb: Fix temporarily to prevent a segmentation fault     under Oracle 9.2.0.1. Fix for BigDecimal bundled in ruby 1.8.5. * test/test_oranumber.rb: Fix for Oracle 9.2. Comparing float values     by <=> is too sensitive to use it in tests.
git-svn-id: svn+ssh://rubyforge.org/var/svn/ruby-oci8/trunk/ruby-oci8@364 f081da96-e817-0410-9439-ba585a54bbf1

+++ test/test_oci8.rb
@@ -167,6 +167,9 @@ EOS
   end
 
   def test_bind_cursor
+    # FIXME: check again after upgrading Oracle 9.2 to 9.2.0.4.
+    return if $oracle_version < OCI8::ORAVER_10_1
+
     drop_table('test_table')
     sql = <<-EOS
 CREATE TABLE test_table
@@ -357,7 +360,7 @@ EOS
       assert_equal(row[0], 12345678901234)
       assert_equal(row[1], 12345678901234567890)
       assert_equal(row[2], 123456789012.34)
-      assert_equal(row[3], 1234567890123.45)
+      assert_equal(row[3], BigDecimal("1234567890123.45"))
       assert_equal(row[4], 1234.5)
       assert_instance_of(BigDecimal, row[0])
       assert_instance_of(Bignum, row[1])
